Add files via upload
The dataset applied for this research offers a detailed analysis of daily technology engagements, such as social media usage and screen time, and mental health. This dataset has been obtained from an online survey conducted in the year 2022, which provides more significance and recent records of the behavioral pattern of people with digital platforms and their influence on wellbeing status. The initial dataset consists of 10,000 records (thousand rows of observation data) and 14 features (14 pieces of column data) in terms of participants’ daily activity and mental state.